+++ package.json
@@ -9,18 +9,18 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "next": "14.2.2"
   },
   "devDependencies": {
-    "typescript": "^5",
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "postcss": "^8",
+    "typescript": "^4.5.4",
+    "@types/node": "^16.11.15",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "postcss": "^8.4.6",
     "tailwindcss": "^3.4.1",
-    "eslint": "^8",
+    "eslint": "^8.5.0",
     "eslint-config-next": "14.2.2"
   }
 }
