+++ pages/index.tsx
@@ -1,0 +1,8 @@
+// pages/index.tsx
+
+import React from 'react';
+import Home from './home';
+
+export default function Index() {
+  return <Home />;
+}
